requirements.txt
emoji~=2.9.0
aiogram~=3.2.0
SQLAlchemy~=2.0.23
python-dotenv~=1.0.0
aiosqlite~=0.19.0